refactor: separate bluepy circuit access def from sonata
