+++ bluecellulab/circuit/circuit_access/__init__.py
@@ -1,0 +1,3 @@
+from .definition import EmodelProperties, CircuitAccess, get_synapse_connection_parameters
+from .bluepy_circuit_access import BluepyCircuitAccess
+from .sonata_circuit_access import SonataCircuitAccess
